requirements.txt
pandas==2.2.3
matplotlib==3.9.4
pywebview==5.4
requests==2.32.3